+++ src/optlang/gurobi_interface.py
@@ -34,6 +34,14 @@ from optlang import symbolics
 
 import gurobipy
 
+try:
+    version = gurobipy.gurobi.version()
+    if not (version[0] >= 9 and version[1] >= 5):
+        raise RuntimeError()
+except Exception:
+    raise RuntimeError(
+        "This version of optlang requires a Gurobi version of 9.5 or above.")
+
 _GUROBI_STATUS_TO_STATUS = {
     gurobipy.GRB.LOADED: interface.LOADED,
     gurobipy.GRB.OPTIMAL: interface.OPTIMAL,
@@ -59,9 +67,6 @@ _REVERSE_QP_METHODS = {v: k for k, v in _QP_METHODS.items()}
 _VTYPE_TO_GUROBI_VTYPE = {'continuous': gurobipy.GRB.CONTINUOUS, 'integer': gurobipy.GRB.INTEGER,
                           'binary': gurobipy.GRB.BINARY}
 _GUROBI_VTYPE_TO_VTYPE = dict((val, key) for key, val in _VTYPE_TO_GUROBI_VTYPE.items())
-
-# hacky way since Gurobi does not allow getting the version via the API
-_IS_GUROBI_9_OR_NEWER = hasattr(gurobipy, "MLinExpr")
 
 
 def _constraint_lb_and_ub_to_gurobi_sense_rhs_and_range_value(lb, ub):
@@ -93,16 +98,11 @@ def _constraint_lb_and_ub_to_gurobi_sense_rhs_and_range_value(lb, ub):
 class Variable(interface.Variable):
     def __init__(self, name, *args, **kwargs):
         super(Variable, self).__init__(name, **kwargs)
-        self._original_name = name
 
     @property
     def _internal_variable(self):
         if getattr(self, 'problem', None) is not None:
-            if _IS_GUROBI_9_OR_NEWER:
-                internal_variable = self.problem.problem.getVarByName(self.name)
-            else:
-                internal_variable = self.problem.problem.getVarByName(self._original_name)
-            return internal_variable
+            return self.problem.problem.getVarByName(self.name)
         else:
             return None
 
@@ -739,6 +739,57 @@ class Model(interface.Model):
         super(Model, self)._remove_constraints(constraints)
         for internal_constraint in internal_constraints:
             self.problem.remove(internal_constraint)
+
+    def _get_variables_names(self):
+        """The names of model variables.
+
+        Returns
+        -------
+        list
+        """
+        return self.problem.VarName
+
+    def _get_constraint_names(self):
+        """The names of model constraints.
+
+        Returns
+        -------
+        list
+        """
+        return self.problem.ConstrName
+
+    def _get_primal_values(self):
+        """The primal values of model variables.
+
+        Returns
+        -------
+        list
+        """
+        return self.problem.X
+
+    def _get_reduced_costs(self):
+        """The reduced costs/dual values of all variables.
+
+        Returns
+        -------
+        list
+        """
+        if self.is_integer:
+            raise ValueError(
+                "Reduced costs are not well defined for integer problems.")
+        return self.problem.RC
+
+    def _get_shadow_prices(self):
+        """The shadow prices of model (dual values of all constraints).
+
+        Returns
+        -------
+        collections.OrderedDict
+        """
+        if self.is_integer:
+            raise ValueError(
+                "Shadow prices are not well defined for integer problems.")
+        return self.problem.Pi
 
     @property
     def is_integer(self):